+++ lib/plausible_web/controllers/site_controller.ex
@@ -179,11 +179,11 @@ defmodule PlausibleWeb.SiteController do
       end
 
     conn
-    |> assign(:skip_plausible_tracking, true)
     |> render("settings_general.html",
       site: site,
       imported_pageviews: imported_pageviews,
       changeset: Plausible.Site.changeset(site, %{}),
+      dogfood_page_path: "/:dashboard/settings/general",
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
   end
@@ -194,9 +194,9 @@ defmodule PlausibleWeb.SiteController do
       |> Repo.preload(memberships: :user, invitations: [], custom_domain: [])
 
     conn
-    |> assign(:skip_plausible_tracking, true)
     |> render("settings_people.html",
       site: site,
+      dogfood_page_path: "/:dashboard/settings/people",
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
   end
@@ -206,10 +206,10 @@ defmodule PlausibleWeb.SiteController do
     shared_links = Repo.all(from l in Plausible.Site.SharedLink, where: l.site_id == ^site.id)
 
     conn
-    |> assign(:skip_plausible_tracking, true)
     |> render("settings_visibility.html",
       site: site,
       shared_links: shared_links,
+      dogfood_page_path: "/:dashboard/settings/visibility",
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
   end
@@ -218,9 +218,9 @@ defmodule PlausibleWeb.SiteController do
     site = conn.assigns[:site] |> Repo.preload(:custom_domain)
 
     conn
-    |> assign(:skip_plausible_tracking, true)
     |> render("settings_goals.html",
       site: site,
+      dogfood_page_path: "/:dashboard/settings/goals",
       connect_live_socket: true,
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
@@ -233,6 +233,7 @@ defmodule PlausibleWeb.SiteController do
     |> assign(:skip_plausible_tracking, true)
     |> render("settings_funnels.html",
       site: site,
+      dogfood_page_path: "/:dashboard/settings/funnels",
       connect_live_socket: true,
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
@@ -245,6 +246,7 @@ defmodule PlausibleWeb.SiteController do
     |> assign(:skip_plausible_tracking, true)
     |> render("settings_props.html",
       site: site,
+      dogfood_page_path: "/:dashboard/settings/properties",
       layout: {PlausibleWeb.LayoutView, "site_settings.html"},
       connect_live_socket: true
     )
@@ -261,10 +263,10 @@ defmodule PlausibleWeb.SiteController do
       end
 
     conn
-    |> assign(:skip_plausible_tracking, true)
     |> render("settings_search_console.html",
       site: site,
       search_console_domains: search_console_domains,
+      dogfood_page_path: "/:dashboard/settings/search-console",
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
   end
@@ -273,12 +275,12 @@ defmodule PlausibleWeb.SiteController do
     site = conn.assigns[:site] |> Repo.preload(:custom_domain)
 
     conn
-    |> assign(:skip_plausible_tracking, true)
     |> render("settings_email_reports.html",
       site: site,
       weekly_report: Repo.get_by(Plausible.Site.WeeklyReport, site_id: site.id),
       monthly_report: Repo.get_by(Plausible.Site.MonthlyReport, site_id: site.id),
       spike_notification: Repo.get_by(Plausible.Site.SpikeNotification, site_id: site.id),
+      dogfood_page_path: "/:dashboard/settings/email-reports",
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
   end
@@ -300,9 +302,9 @@ defmodule PlausibleWeb.SiteController do
     site = conn.assigns[:site] |> Repo.preload(:custom_domain)
 
     conn
-    |> assign(:skip_plausible_tracking, true)
     |> render("settings_danger_zone.html",
       site: site,
+      dogfood_page_path: "/:dashboard/settings/danger-zone",
       layout: {PlausibleWeb.LayoutView, "site_settings.html"}
     )
   end